Convert .txt LCD output files to HDF5 format
Given a .txt input file that is an output file (converted from .root
format) from the DDhep LCD simulation framework, the cript calculates
the centoid of the hit registered on the calorimeters. Mapping from
local to global coordinates, the centroid calculated in the
Electromagnetic Calorimeter is subsequently mapped to the corresponding
cells of the Hadronic Calorimeter.

A regular array of 25x25x25 cells around, and including, the centroid
are taken to form the image of the shower in the ECAL, similarly a
4x4x60 array is made for the HCAL data.